Add - View Booking

+++ lib/helpers.py
@@ -14,7 +14,7 @@ def clear_screen():
     os.system("clear")
 
 
-def main_menu():
+def intro():
     print(
         "============================================================================"
     )
@@ -28,44 +28,56 @@ def main_menu():
 def menu_option():
     print("Please choose what you want to do")
     print("Input 1 - View booking details")
-    print("Input 2 - Make a new booking")
-    print("Input 3 - Update a booking")
-    print("Input 4 - Delete a booking")
-    print("Input 5 - Quit")
+    print("Input 2 - Quit")
 
 
-def prompt_input():
-    menu_option()
+def submenu_option():
+    print("Input 1 - Make a new booking")
+    print("Input 2 - Update a booking")
+    print("Input 3 - Delete a booking")
+    print("Input 4 - Quit")
+
+
+def prompt_input(options):
+    if options == 2:
+        menu_option()
+    elif options == 4:
+        submenu_option()
+
     choice = int(input("Please Enter your choice: "))
-    if choice <= 4:
+    if choice <= options:
         return choice
     else:
-        print("Please input correct range number from 1 to 4")
+        print("Please input correct range number")
         return None
 
 
-def user_request():
-    selected_value = prompt_input()
+def user_request(options):
+    selected_value = prompt_input(options)
 
     while not selected_value:
-        selected_value = prompt_input()
+        selected_value = prompt_input(options)
+    if selected_value:
+        return selected_value
 
-    option_action(selected_value)
 
-
-def option_action(value):
+def menu_action(value):
     clear_screen()
     if value == 1:
         view_booking()
     elif value == 2:
-        add_new_booking()
-    elif value == 3:
-        update_booking()
-    elif value == 4:
-        delete_booking()
-    elif value == 5:
         print("Quit the application successfully!")
         exit()
+
+
+def submenu_action(value):
+    clear_screen()
+    if value == 1:
+        add_new_booking()
+    elif value == 2:
+        update_booking()
+    elif value == 3:
+        delete_booking()
 
 
 def prompt_restaurant():
@@ -88,6 +100,11 @@ def prompt_email():
     return email
 
 
+def email_not_exist(email):
+    print(f"Sorry, the {email} does not exist")
+    print("Do you wish to input another email?")
+
+
 def view_booking():
     clear_screen()
     email = prompt_email()
@@ -96,20 +113,23 @@ def view_booking():
 
     if user:
         print(f"User Name: {user.name}")
-        for booking in user.bookings:
-            print(f"Booking Time: ${booking.time}")
-            print(f"Booking Restaurant: ${booking.restaurant.name}")
-        user_request()
+        for index, booking in enumerate(user.bookings):
+            print(f"=========== {index+1} ===========")
+            print(f"Booking Date: {booking.date}")
+            print(f"Booking Time: {booking.time}")
+            print(f"Booking Restaurant: {booking.restaurant.name}")
+            print("=================================")
+            choice = user_request(4)
+            submenu_action(choice)
     else:
-        print(f"Sorry, the {email} does not exist")
-        print("Do you wish to input another email?")
+        email_not_exist(email)
         choice = int(
             input("Input 1 to attempt your email again, Input 2 go back to home page")
         )
         if choice == 1:
             view_booking()
         elif choice == 2:
-            user_request()
+            user_request(2)
 
 
 def add_new_booking():
@@ -131,7 +151,7 @@ def add_new_booking():
 
     session.add(new_user)
     session.commit()
-    user_request()
+    user_request(4)
 
 
 def update_booking():
@@ -147,19 +167,30 @@ def update_booking():
         else:
             print("Sorry We can not find your bookings")
             update_booking()
-    user_request()
+    user_request(4)
 
 
 def delete_booking():
-    email = input("Please Input your booking email=> ")
+    clear_screen()
+    email = prompt_email()
     user = session.query(User).filter(User.email == email).first()
-    for index, booking in enumerate(user.bookings):
-        print(f"Input your {index} number to update your {booking}")
-        choice = int(input("Please Enter your choice: "))
-        if index == choice:
-            session.delete(booking)
-            session.commit()
-        else:
-            print("Sorry We can not find your bookings")
+
+    if user:
+        for index, booking in enumerate(user.bookinbgs):
+            print(
+                f"Input your {index} number to update your ${booking.restaurant.name}, Booking Time: ${booking.time}"
+            )
+            choice = int(input("Please Enter your choice: "))
+            if index == choice:
+                session.delete(booking)
+                session.commit()
+        user_request(4)
+    else:
+        email_not_exist(email)
+        choice = int(
+            input("Input 1 to attempt your email again, Input 2 go back to home page")
+        )
+        if choice == 1:
             delete_booking()
-    user_request()
+        elif choice == 2:
+            user_request(2)
